+++ node/src/types/manifest/index.d.ts
@@ -23,4 +23,5 @@ export * as base from "./base";
 export * as icon from "./icon";
 export * as plugin from "./plugin";
 export * as template from "./template";
+export * as widget from "./widget";
 //#endregion content