+++ node/src/types/manifest/index.d.ts
@@ -23,5 +23,6 @@ export * as base from "./base";
 export * as icon from "./icon";
 export * as plugin from "./plugin";
 export * as template from "./template";
+export * as theme from "./theme";
 export * as widget from "./widget";
 //#endregion content