+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 添加 API `/api/export/exportHTML` | Add API `/api/export/exportHTML`
 - 调整 API `/api/export/exportHTML` | Adjust API `/api/export/exportHTML`
+- 优化 WebSocket URL 生成方案 | Optimize WebSocket URL generation scheme
 
 ## v0.3.9 / 2024-04-06