+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - 调整 API `/api/export/exportHTML` | Adjust API `/api/export/exportHTML`
 - 优化 WebSocket URL 生成方案 | Optimize WebSocket URL generation scheme
 - 调整 API `/api/system/getConf` | Adjust API `/api/system/getConf`
+- 调整 API `/api/notebook/getNotebookConf` | Adjust API `/api/notebook/getNotebookConf`
 
 ## v0.3.9 / 2024-04-06