+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 更改日志 | Change Log
 
+## v0.3.10 / 2024-05-09
+
+- [v0.3.9 ... v0.3.10](https://github.com/siyuan-community/siyuan-sdk/compare/v0.3.9...v0.3.10)
 - 添加 API `/api/export/exportHTML` | Add API `/api/export/exportHTML`
 - 调整 API `/api/export/exportHTML` | Adjust API `/api/export/exportHTML`
 - 优化 WebSocket URL 生成方案 | Optimize WebSocket URL generation scheme